Restructure token tests
