--- contracts/contracts/cairo/ocr2/token.cairo
@@ -1,3 +1,0 @@
-%lang starknet
-
-from openzeppelin.token.erc20.presets.ERC20Mintable import constructor
